feat: converter 파일 분리 및 내부 모델에서 toJson을 재정의
